chg: organise choices and range code into validators module
